+++ src/Concerto/TestBundle/Command/StartForkerCommand.php
@@ -35,10 +35,10 @@ class StartForkerCommand extends Command
     {
         $forkerPath = realpath(dirname(__FILE__) . "/../Resources/R/forker.R");
         $logPath = realpath(dirname(__FILE__) . "/../Resources/R") . "/forker.log";
-        $fifoPath = realpath(dirname(__FILE__) . "/../Resources/R/fifo") . "/";
-        $publicDir = realpath(dirname(__FILE__) . "/../../PanelBundle/Resources/public/files") . "/";
+        $fifoPath = $this->sessionRunnerService->getFifoDir();
+        $publicDir = $this->sessionRunnerService->getPublicDirPath();
         $connection = json_encode($this->sessionRunnerService->getConnection());
-        $mediaUrl = $this->testRunnerSettings["dir"] . "bundles/concertopanel/files/";
+        $platformUrl = $this->sessionRunnerService->getPlatformUrl();
         $maxExecTime = $this->testRunnerSettings["max_execution_time"];
         $maxIdleTime = $this->testRunnerSettings["max_idle_time"];
         $keepAliveToleranceTime = $this->testRunnerSettings["keep_alive_tolerance_time"];
@@ -47,7 +47,7 @@ class StartForkerCommand extends Command
             . "'$forkerPath' "
             . "'$fifoPath' "
             . "'$publicDir' "
-            . "'$mediaUrl' "
+            . "'$platformUrl' "
             . "'$connection' "
             . "$maxExecTime "
             . "$maxIdleTime "