+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.file.getPath.R
@@ -1,3 +1,3 @@
-concerto.file.getPublicPath = function(filename){
+concerto.file.getPath = function(filename){
     return(paste0(concerto$publicDir, filename))
 }