+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.saml.login.R
@@ -1,0 +1,5 @@
+concerto.saml.login = function(redirectTo=NULL){
+    if(is.null(redirectTo)) redirectTo = concerto.session.getResumeUrl()
+    url = paste0(concerto$platformUrl, "api/saml/login?redirectTo=", redirectTo)
+    concerto.template.redirect(url)
+}