+++ src/Concerto/TestBundle/Resources/R/concerto5/R/concerto.template.redirect.R
@@ -1,0 +1,6 @@
+concerto.template.redirect = function(url) {
+    concerto.template.show(
+        html=paste0("<script>location.href='",url,"'</script>"),
+        skipOnResume=T
+    )
+}