+++ src/Concerto/TestBundle/Service/SerializedSessionRunnerService.php
@@ -16,9 +16,10 @@ class SerializedSessionRunnerService extends ASessionRunnerService
 
     public function __construct(LoggerInterface $logger, TestSessionRepository $testSessionRepository, AdministrationService $administrationService, TestSessionCountService $testSessionCountService, RegistryInterface $doctrine, $testRunnerSettings, $root, $environment)
     {
-        parent::__construct($logger, $testRunnerSettings, $root, $doctrine, $testSessionCountService, $administrationService, $testSessionRepository);
-
+        $this->runnerType = 1;
         $this->environment = $environment;
+
+        parent::__construct($logger, $testRunnerSettings, $root, $doctrine, $testSessionCountService, $administrationService, $testSessionRepository);
     }
 
     public function healthCheck()
@@ -41,16 +42,16 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
 
         $response = $this->startListenerSocket($submitter_sock);
+        socket_close($submitter_sock);
         if ($response === false) {
             return false;
         }
         $response = json_decode($response, true);
-        socket_close($submitter_sock);
 
         return $response["code"] === 1;
     }
 
-    public function startNew(TestSession $session, $params, $client_ip, $client_browser, $debug = false, $max_exec_time = null)
+    public function startNew(TestSession $session, $params, $cookies, $client_ip, $client_browser, $debug = false, $max_exec_time = null)
     {
         $session_hash = $session->getHash();
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $params, $client_ip, $client_ip, $client_browser, $debug");
@@ -64,10 +65,17 @@ class SerializedSessionRunnerService extends ASessionRunnerService
 
         if (!$this->createSubmitterSock($session, false, $submitter_sock, $error_response)) return $error_response;
 
+        $request = array(
+            "source" => TestSessionService::SOURCE_PANEL_NODE,
+            "code" => TestSessionService::RESPONSE_STARTING,
+            "client" => $client,
+            "cookies" => $cookies
+        );
+
         if ($this->getOS() == self::OS_LINUX && $this->testRunnerSettings["session_forking"] == "true") {
-            $success = $this->startChildProcess($client, $session_hash,null, $max_exec_time);
+            $success = $this->startChildProcess($client, $session_hash, $request, $max_exec_time);
         } else {
-            $success = $this->startStandaloneProcess($client, $session_hash, null, $max_exec_time);
+            $success = $this->startStandaloneProcess($client, $session_hash, $request, $max_exec_time);
         }
         if (!$success) {
             socket_close($submitter_sock);
@@ -79,6 +87,7 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
 
         $response = $this->startListenerSocket($submitter_sock, $max_exec_time);
+        socket_close($submitter_sock);
         if ($response === false) {
             return array(
                 "source" => TestSessionService::SOURCE_TEST_NODE,
@@ -87,13 +96,62 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
         $response = json_decode($response, true);
         $response = $this->appendDebugDataToResponse($session, $response);
-        socket_close($submitter_sock);
 
         $this->testSessionRepository->clear();
         return $response;
     }
 
-    public function submit(TestSession $session, $values, $client_ip, $client_browser)
+    public function resume(TestSession $session, $cookies, $client_ip, $client_browser, $debug = false, $max_exec_time = null)
+    {
+        $session_hash = $session->getHash();
+        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_ip, $client_browser, $debug");
+
+        if (!$this->checkSessionLimit($session, $response)) return $response;
+
+        $client = array(
+            "ip" => $client_ip,
+            "browser" => $client_browser
+        );
+
+        if (!$this->createSubmitterSock($session, false, $submitter_sock, $error_response)) return $error_response;
+
+        $request = array(
+            "source" => TestSessionService::SOURCE_PANEL_NODE,
+            "code" => TestSessionService::RESPONSE_RESUME,
+            "client" => $client,
+            "cookies" => $cookies
+        );
+
+        if ($this->getOS() == self::OS_LINUX && $this->testRunnerSettings["session_forking"] == "true") {
+            $success = $this->startChildProcess($client, $session_hash, $request, $max_exec_time);
+        } else {
+            $success = $this->startStandaloneProcess($client, $session_hash, $request, $max_exec_time);
+        }
+        if (!$success) {
+            socket_close($submitter_sock);
+            $this->logger->error(__CLASS__ . ":" . __FUNCTION__ . " - creating R process failed");
+            return array(
+                "source" => TestSessionService::SOURCE_TEST_NODE,
+                "code" => TestSessionService::RESPONSE_ERROR
+            );
+        }
+
+        $response = $this->startListenerSocket($submitter_sock, $max_exec_time);
+        socket_close($submitter_sock);
+        if ($response === false) {
+            return array(
+                "source" => TestSessionService::SOURCE_TEST_NODE,
+                "code" => TestSessionService::RESPONSE_ERROR
+            );
+        }
+        $response = json_decode($response, true);
+        $response = $this->appendDebugDataToResponse($session, $response);
+
+        $this->testSessionRepository->clear();
+        return $response;
+    }
+
+    public function submit(TestSession $session, $values, $cookies, $client_ip, $client_browser)
     {
         $session_hash = $session->getHash();
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
@@ -106,17 +164,18 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         if (!$this->createSubmitterSock($session, true, $submitter_sock, $error_response)) return $error_response;
         $debugOffset = $this->getDebugDataOffset($session);
 
-        $response = array(
+        $request = array(
             "source" => TestSessionService::SOURCE_PANEL_NODE,
             "code" => TestSessionService::RESPONSE_SUBMIT,
             "client" => $client,
-            "values" => $values
+            "values" => $values,
+            "cookies" => $cookies
         );
 
         if ($this->getOS() == self::OS_LINUX && $this->testRunnerSettings["session_forking"] == "true") {
-            $success = $this->startChildProcess($client, $session_hash, $response);
+            $success = $this->startChildProcess($client, $session_hash, $request);
         } else {
-            $success = $this->startStandaloneProcess($client, $session_hash, $response);
+            $success = $this->startStandaloneProcess($client, $session_hash, $request);
         }
         if (!$success) {
             socket_close($submitter_sock);
@@ -128,6 +187,7 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
 
         $response = $this->startListenerSocket($submitter_sock);
+        socket_close($submitter_sock);
         if ($response === false) {
             return array(
                 "source" => TestSessionService::SOURCE_TEST_NODE,
@@ -136,13 +196,12 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
         $response = json_decode($response, true);
         $response = $this->appendDebugDataToResponse($session, $response, $debugOffset);
-        socket_close($submitter_sock);
 
         $this->testSessionRepository->clear();
         return $response;
     }
 
-    public function backgroundWorker(TestSession $session, $values, $client_ip, $client_browser)
+    public function backgroundWorker(TestSession $session, $values, $cookies, $client_ip, $client_browser)
     {
         $session_hash = $session->getHash();
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $values, $client_ip, $client_browser");
@@ -159,7 +218,8 @@ class SerializedSessionRunnerService extends ASessionRunnerService
             "source" => TestSessionService::SOURCE_PANEL_NODE,
             "code" => TestSessionService::RESPONSE_WORKER,
             "client" => $client,
-            "values" => $values
+            "values" => $values,
+            "cookies" => $cookies
         );
 
         if ($this->getOS() == self::OS_LINUX && $this->testRunnerSettings["session_forking"] == "true") {
@@ -177,6 +237,7 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
 
         $response = $this->startListenerSocket($submitter_sock);
+        socket_close($submitter_sock);
         if ($response === false) {
             return array(
                 "source" => TestSessionService::SOURCE_TEST_NODE,
@@ -185,7 +246,6 @@ class SerializedSessionRunnerService extends ASessionRunnerService
         }
         $response = json_decode($response, true);
         $response = $this->appendDebugDataToResponse($session, $response, $debugOffset);
-        socket_close($submitter_sock);
 
         $this->testSessionRepository->clear();
         return $response;
@@ -211,122 +271,5 @@ class SerializedSessionRunnerService extends ASessionRunnerService
             "source" => TestSessionService::SOURCE_PROCESS,
             "code" => TestSessionService::RESPONSE_STOPPED
         );
-    }
-
-    private function startStandaloneProcess($client, $session_hash, $response = null, $max_exec_time = null, $initial_port = null)
-    {
-        $cmd = $this->getCommand($client, $session_hash, $response, $max_exec_time, $initial_port);
-        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $cmd");
-
-        $process = new Process($cmd);
-        $process->setEnhanceWindowsCompatibility(false);
-
-        $env = array(
-            "R_GC_MEM_GROW" => 0
-        );
-        if ($this->testRunnerSettings["r_environ_path"] != null) {
-            $env["R_ENVIRON"] = $this->testRunnerSettings["r_environ_path"];
-        }
-        $process->setEnv($env);
-        $process->mustRun();
-        return true;
-    }
-
-    private function startChildProcess($client, $session_hash, $response = null, $max_exec_time = null, $initial_port = null)
-    {
-        if ($max_exec_time === null) {
-            $max_exec_time = $this->testRunnerSettings["max_execution_time"];
-        }
-
-        $rout = $this->getROutputFilePath($session_hash);
-
-        $response = json_encode(array(
-            "workingDir" => realpath($this->getWorkingDirPath($session_hash)) . "/",
-            "maxExecTime" => $max_exec_time,
-            "maxIdleTime" => $this->administrationService->getSettingValueForSessionHash($session_hash, "max_idle_time"),
-            "keepAliveToleranceTime" => $this->testRunnerSettings["keep_alive_tolerance_time"],
-            "client" => $client,
-            "connection" => $this->getConnection(),
-            "sessionId" => $session_hash,
-            "rLogPath" => $rout,
-            "response" => $response,
-            "initialPort" => $initial_port,
-            "runnerType" => 1
-        ));
-
-        $path = $this->getFifoDir() . ($session_hash ? $session_hash : uniqid("hc", true)) . ".fifo";
-        posix_mkfifo($path, POSIX_S_IFIFO | 0644);
-        $fh = fopen($path, "wt");
-        if ($fh === false) {
-            $this->logger->error(__CLASS__ . ":" . __FUNCTION__ . " - fopen() failed");
-            return false;
-        }
-        stream_set_blocking($fh, 1);
-        $buffer = $response . "\n";
-        $sent = fwrite($fh, $buffer);
-        $success = $sent !== false;
-        if (!$success) {
-            $this->logger->error(__CLASS__ . ":" . __FUNCTION__ . " - fwrite() failed");
-        }
-        if (strlen($buffer) != $sent) {
-            $this->logger->error(__CLASS__ . ":" . __FUNCTION__ . " - fwrite() failed, sent only $sent/" . strlen($buffer));
-            $success = false;
-        }
-        fclose($fh);
-        return $success;
-    }
-
-    private function getCommand($client, $session_hash, $response, $max_exec_time, $initial_port)
-    {
-        $rscript_exec = $this->testRunnerSettings["rscript_exec"];
-        $ini_path = $this->getRDir() . "standalone.R";
-        $max_exec_time = $max_exec_time === null ? $this->testRunnerSettings["max_execution_time"] : $max_exec_time;
-        $max_idle_time = $this->administrationService->getSettingValueForSessionHash($session_hash, "max_idle_time");
-        $keep_alive_tolerance_time = $this->testRunnerSettings["keep_alive_tolerance_time"];
-        $database_connection = json_encode($this->getConnection());
-        $working_directory = $this->getWorkingDirPath($session_hash);
-        $public_directory = $this->getPublicDirPath();
-        $media_url = $this->getMediaUrl();
-        $client = json_encode($client);
-        $response = json_encode($response ? $response : array());
-        $rout = $this->getROutputFilePath($session_hash);
-
-        switch ($this->getOS()) {
-            case self::OS_LINUX:
-                return "nohup " . $rscript_exec . " --no-save --no-restore --quiet "
-                    . "'$ini_path' "
-                    . "'$database_connection' "
-                    . "'$client' "
-                    . $session_hash . " "
-                    . "'$working_directory' "
-                    . "'$public_directory' "
-                    . "'$media_url' "
-                    . "$max_exec_time "
-                    . "$max_idle_time "
-                    . "$keep_alive_tolerance_time "
-                    . "'$response' "
-                    . "$initial_port "
-                    . "1 " //runnerType
-                    . ($rout ? (">> '" . $rout . "' ") : "")
-                    . "2>&1 & echo $!";
-            default:
-                return "start cmd /C \""
-                    . "\"" . $this->escapeWindowsArg($rscript_exec) . "\" --no-save --no-restore --quiet "
-                    . "\"" . $this->escapeWindowsArg($ini_path) . "\" "
-                    . "\"" . $this->escapeWindowsArg($database_connection) . "\" "
-                    . "\"" . $this->escapeWindowsArg($client) . "\" "
-                    . $session_hash . " "
-                    . "\"" . $this->escapeWindowsArg($working_directory) . "\" "
-                    . "\"" . $this->escapeWindowsArg($public_directory) . "\" "
-                    . "$media_url "
-                    . "$max_exec_time "
-                    . "$max_idle_time "
-                    . "$keep_alive_tolerance_time "
-                    . "\"" . $this->escapeWindowsArg($response) . "\" "
-                    . "$initial_port "
-                    . "1 " //runnerType
-                    . ($rout ? (">> \"" . $this->escapeWindowsArg($rout) . "\" ") : "")
-                    . "2>&1\"";
-        }
     }
 }