+++ src/Concerto/TestBundle/Service/TestRunnerService.php
@@ -17,29 +17,37 @@ class TestRunnerService {
         $this->sessionService = $sessionService;
     }
 
-    public function startNewSession($test_slug, $test_name, $params, $client_ip, $client_browser, $debug) {
+    public function startNewSession($test_slug, $test_name, $params, $cookies, $client_ip, $client_browser, $debug) {
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $test_slug, $test_name, $params, $client_ip, $client_browser, $debug");
 
-        $response = $this->sessionService->startNewSession($test_slug, $test_name, $params, $client_ip, $client_browser, $debug);
+        $response = $this->sessionService->startNewSession($test_slug, $test_name, $params, $cookies, $client_ip, $client_browser, $debug);
         $response = json_encode($response);
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
         return $response;
     }
 
-    public function submitToSession($session_hash, $values, $client_ip, $client_browser) {
-        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
+    public function resumeSession($session_hash, $cookies, $client_ip, $client_browser, $debug) {
+        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser, $debug");
 
-
-        $response = $this->sessionService->submit($session_hash, $values, $client_ip, $client_browser);
+        $response = $this->sessionService->resumeSession($session_hash, $cookies, $client_ip, $client_browser, $debug);
         $response = json_encode($response);
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
         return $response;
     }
 
-    public function backgroundWorker($session_hash, $values, $client_ip, $client_browser) {
+    public function submitToSession($session_hash, $values, $cookies, $client_ip, $client_browser) {
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
 
-        $response = $this->sessionService->backgroundWorker($session_hash, $values, $client_ip, $client_browser);
+        $response = $this->sessionService->submit($session_hash, $values, $cookies, $client_ip, $client_browser);
+        $response = json_encode($response);
+        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
+        return $response;
+    }
+
+    public function backgroundWorker($session_hash, $values, $cookies, $client_ip, $client_browser) {
+        $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - $session_hash, $client_ip, $client_browser");
+
+        $response = $this->sessionService->backgroundWorker($session_hash, $values, $cookies, $client_ip, $client_browser);
         $response = json_encode($response);
         $this->logger->info(__CLASS__ . ":" . __FUNCTION__ . " - response: $response");
         return $response;